added checks product not found error

+++ origami/controllers/product.ts
@@ -1,7 +1,6 @@
 import { Request, ResponseObject, ResponseToolkit } from '@hapi/hapi';
-import { logger, handleDBError, prepareResponse } from '@utilities';
+import { logger, handleDBError, prepareResponse, productNotFoundError } from '@utilities';
 import { ProductCreationPayload, ProductUpdationPayload } from '@interfaces/product';
-import { productNotFoundError } from '@utilities';
 
 // eslint-disable-next-line import/prefer-default-export
 export const createProduct = async (
@@ -13,19 +12,21 @@ export const createProduct = async (
   const { traceId } = request.pre;
   try {
     const productInfo = {
-      ...payload
-    }
-    const product = await productService.createProduct({ productInfo })
-    return h.response(prepareResponse({
-      success: true, data: {
-        product
-      }
-    }));
+      ...payload,
+    };
+    const product = await productService.createProduct({ productInfo });
+    return h
+      .response(
+        prepareResponse({
+          success: true,
+          data: {
+            product,
+          },
+        }),
+      )
+      .code(201);
   } catch (err) {
-    logger.error(
-      `traceId=${traceId} createProduct error=${err.message}`,
-      err,
-    );
+    logger.error(`traceId=${traceId} createProduct error=${err.message}`, err);
     throw handleDBError({ err });
   }
 };
@@ -39,18 +40,22 @@ export const updateProduct = async (
   const { productService } = request.services();
   const { traceId } = request.pre;
   try {
-    const productInfo = {
-      ...payload
+    const product = await productService.findProductById({ id: Number(productId) });
+    if (!product) {
+      throw productNotFoundError(productId);
     }
-    await productService.updateProduct({ id: Number(productId), productInfo })
-    return h.response(prepareResponse({
-      success: true, data: {}
-    }));
-  } catch (err) {
-    logger.error(
-      `traceId=${traceId} updateProduct error=${err.message}`,
-      err,
+    const productInfo = {
+      ...payload,
+    };
+    await productService.updateProduct({ id: Number(productId), productInfo });
+    return h.response(
+      prepareResponse({
+        success: true,
+        data: {},
+      }),
     );
+  } catch (err) {
+    logger.error(`traceId=${traceId} updateProduct error=${err.message}`, err);
     throw handleDBError({ err });
   }
 };
@@ -59,7 +64,6 @@ export const getProductById = async (
   request: Request,
   h: ResponseToolkit,
 ): Promise<ResponseObject> => {
-
   const { id: productId } = request.params;
   const { productService } = request.services();
   const { traceId } = request.pre;
@@ -68,16 +72,16 @@ export const getProductById = async (
     if (!product) {
       throw productNotFoundError(productId);
     }
-    return h.response(prepareResponse({
-      success: true, data: {
-        product,
-      }
-    }));
-  } catch (err) {
-    logger.error(
-      `traceId=${traceId} getProductById error=${err.message}`,
-      err,
+    return h.response(
+      prepareResponse({
+        success: true,
+        data: {
+          product,
+        },
+      }),
     );
+  } catch (err) {
+    logger.error(`traceId=${traceId} getProductById error=${err.message}`, err);
     throw handleDBError({ err });
   }
 };
@@ -86,20 +90,23 @@ export const deleteProductById = async (
   request: Request,
   h: ResponseToolkit,
 ): Promise<ResponseObject> => {
-
   const { id: productId } = request.params;
   const { productService } = request.services();
   const { traceId } = request.pre;
   try {
+    const product = await productService.findProductById({ id: Number(productId) });
+    if (!product) {
+      throw productNotFoundError(productId);
+    }
     await productService.deleteProductById({ id: Number(productId) });
-    return h.response(prepareResponse({
-      success: true, data: {}
-    }));
-  } catch (err) {
-    logger.error(
-      `traceId=${traceId} deleteProductById error=${err.message}`,
-      err,
+    return h.response(
+      prepareResponse({
+        success: true,
+        data: {},
+      }),
     );
+  } catch (err) {
+    logger.error(`traceId=${traceId} deleteProductById error=${err.message}`, err);
     throw handleDBError({ err });
   }
 };
@@ -108,20 +115,19 @@ export const getProducts = async (
   request: Request,
   h: ResponseToolkit,
 ): Promise<ResponseObject> => {
-
   const { category, productName, page, limit } = request.query;
   const { productService } = request.services();
   const { traceId } = request.pre;
   try {
     const data = await productService.findByFields({ category, productName, page, limit });
-    return h.response(prepareResponse({
-      success: true, data,
-    }));
-  } catch (err) {
-    logger.error(
-      `traceId=${traceId} deleteProductById error=${err.message}`,
-      err,
+    return h.response(
+      prepareResponse({
+        success: true,
+        data,
+      }),
     );
+  } catch (err) {
+    logger.error(`traceId=${traceId} deleteProductById error=${err.message}`, err);
     throw handleDBError({ err });
   }
 };
